+++ client/src/layers/ListView/index.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useMemo, useState } from 'react';
-import Styled from '@/components/ListViewLayer/ListView/style';
+import Styled from '@/layers/ListView/style';
 import { useUserState } from '@/lib/hooks/useContextHooks';
-import ListViewHeader from '@/components/ListViewLayer/ListViewHeader';
-import ListViewItem from '@/components/ListViewLayer/ListViewItem';
+import ListViewHeader from '@/components/ListViewHeader';
+import ListViewItem from '@/components/ListViewItem';
 import { PrivateTask, ProjectType } from '@/contexts/userContext';
 
 export type ListState = 'all' | 'private' | 'project' | 'done';
